requirements.txt
albumentations==1.1.0
beautifulsoup4==4.11.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
filelock==3.7.1
flake8==4.0.1
flatbuffers==2.0
gdown==4.5.1
idna==3.3
imageio==2.16.1
importlib-metadata==4.2.0
joblib==1.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.5
onnxruntime==1.10.0
onnxruntime-gpu==1.10.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==3.19.4
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.7
PySocks==1.7.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
